requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
aiosmtplib==4.0.0
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.1.0
bcrypt==4.0.1
bidict==0.23.1
billiard==4.2.1
black==25.1.0
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
izulu==0.5.4
Jinja2==3.1.6
kombu==5.4.2
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
polyline==2.0.2
prompt_toolkit==3.0.50
propcache==0.3.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pycron==3.1.2
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pytest==8.3.5
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-jose==3.4.0
python-multipart==0.0.20
python-socketio==5.12.1
pytz==2025.1
redis==5.2.1
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.38
starlette==0.46.0
taskiq==0.11.14
taskiq-dependencies==1.5.7
taskiq-fastapi==0.3.4
taskiq-redis==1.0.3
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.18.3
zipp==3.21.0
